+++ src/app/services/firebase.service.ts
@@ -5,11 +5,11 @@ import { AngularFireStorage } from "@angular/fire/compat/storage";
 import { Router } from "@angular/router";
 import { User } from "../models/user.model";
 import {
-  getStorage,
-  ref,
-  uploadString,
-  uploadBytes,
-  deleteObject,
+  getStorage as getStorageLegacy,
+  ref as refLegacy,
+  uploadString as uploadStringLegacy,
+  uploadBytes as uploadBytesLegacy,
+  deleteObject as deleteObjectLegacy,
 } from "firebase/storage";
 import { getAuth, updatePassword, User as FirebaseUser } from "firebase/auth";
 import { Observable, of } from "rxjs";
@@ -31,8 +31,16 @@ import {
   query,
   updateDoc,
 } from "@angular/fire/firestore";
+import {
+  getDownloadURL,
+  getStorage,
+  ref,
+  uploadString,
+} from "@angular/fire/storage";
 import { WtUser } from "../models/wt-user";
 import { environment } from "src/environments/environment";
+import { Photo } from "./camera.service";
+import { ImageExtension } from "../types/image-extension.type";
 
 /**
  * @todo @diana Esta clase contiene dependencias a módulos de angular fire en modo compat. Esto
@@ -208,39 +216,50 @@ export class FirebaseService {
   }
 
   //============Subir imagenes=================
+  /**
+   * @deprecated This method uses legacy imports. Use uploadFileToStorage or uploadStringToStorage
+   * instead.
+   */
   async uploadPhoto(id, file): Promise<any> {
-    const storage = getStorage();
-    const storageRef = ref(storage, id);
-    return uploadString(storageRef, file, "data_url").then((res) => {
+    const storage = getStorageLegacy();
+    const storageRef = refLegacy(storage, id);
+    return uploadStringLegacy(storageRef, file, "data_url").then((res) => {
       return this.storage.ref(id).getDownloadURL().toPromise();
     });
   }
 
   //============Subir Archivos=================
+  /**
+   * @deprecated This method uses legacy imports. Use uploadFileToStorage or uploadStringToStorage
+   * instead.
+   */
   async uploadBlobFile(id, file): Promise<any> {
-    const storage = getStorage();
-    const storageRef = ref(storage, id);
+    const storage = getStorageLegacy();
+    const storageRef = refLegacy(storage, id);
 
     // 'file' comes from the Blob or File API
-    return uploadBytes(storageRef, file).then((snapshot) => {
+    return uploadBytesLegacy(storageRef, file).then((snapshot) => {
       console.log("Uploaded a blob or file!");
       return this.storage.ref(id).getDownloadURL().toPromise();
     });
   }
 
   //============Eliminar de FireStorage=================
+  /**
+   * @deprecated This method uses legacy imports. Use uploadFileToStorage or uploadStringToStorage instead.
+   */
   deleteFromStorage(path: string) {
-    const storage = getStorage();
-    const desertRef = ref(storage, path);
+    const storage = getStorageLegacy();
+    const desertRef = refLegacy(storage, path);
 
-    return deleteObject(desertRef);
+    return deleteObjectLegacy(desertRef);
   }
 
   // =========Cerrar Sesión===========
   /* Cierra sesión y borra datos almacenados en localstorage. */
 
   /**
-   * @deprecated use UserService.logout instead.
+   * @deprecated use UserService.signOut instead.
    */
   async logout() {
     await this.auth.signOut();
@@ -252,7 +271,7 @@ export class FirebaseService {
 
   /**
    * Sign out user and deletes local storage data.
-   * @deprecated use UserService.logout instead.
+   * @deprecated use UserService.signOut instead.
    */
   async signOut(): Promise<void> {
     await this.auth.signOut();
@@ -374,7 +393,7 @@ export class FirebaseService {
    * @param docPath path del documento (colección/id).
    * @param data Partial con los campos a actualizar.
    * @returns
-   * @throws Failure
+   * @throws {FirestoreFailure} if update fails.
    */
   public async update<T>(docPath: string, data: Partial<T>): Promise<void> {
     return updateDoc<DocumentData>(doc(this.firestore, docPath), data).catch(
@@ -382,8 +401,9 @@ export class FirebaseService {
         const f = FailureUtils.errorToFailure(e);
         if (!environment.production) {
           console.groupCollapsed(
-            `🧰 Firebase Service update [${docPath}] [error]`
+            `🧰 FirebaseService.update [${docPath}] [error]`
           );
+          console.log(f);
           console.log(data);
           console.groupEnd();
         }
@@ -404,5 +424,79 @@ export class FirebaseService {
     data: T | Partial<T>
   ): Promise<DocumentReference<DocumentData>> {
     return addDoc<DocumentData>(collection(this.firestore, colPath), data);
+  }
+
+  /**
+   * Uploads a file to Firebase Storage and returns it's download URL.
+   *
+   * @param photo {Photo} with the dataUrl to upload.
+   * @param folder {String} Path with the folder name.
+   * @param fileName {String} Name of the file with or without extension.
+   * @returns {String} with the download URL.
+   * @throws {StorageFailure} if upload fails.
+   */
+  public async uploadStringToStorage(
+    photo: Photo,
+    folder: string,
+    fileName: string
+  ): Promise<string> {
+    const folderWithoutInitialTrailingSlash = folder
+      .replace(/^\/+/g, "")
+      .replace(/\/+$/, "");
+    const fileNameWithoutExtension = this.fileNameFromPath(fileName);
+    const fileExt = photo.format;
+
+    const fileRef = ref(
+      getStorage(),
+      `${folderWithoutInitialTrailingSlash}/${fileNameWithoutExtension}.${fileExt}`
+    );
+    await uploadString(fileRef, photo.dataUrl, "data_url").catch(
+      (e: unknown) => {
+        const f = FailureUtils.errorToFailure(e);
+        if (!environment.production) {
+          console.groupCollapsed(`🧰 FirebaseService.uploadString [error]`);
+          console.log(f);
+          console.log({ photo, folder, fileName });
+          console.groupEnd();
+        }
+        throw f;
+      }
+    );
+    return getDownloadURL(fileRef);
+  }
+
+  /**
+   * Agrega un prefijo alfanumérico para hacer único el nombre del archivo.
+   *
+   * @param fileName Nombre del archivo
+   * @returns {string}
+   */
+  public static generateUniqueFilename(fileName: string): string {
+    return (
+      Math.random().toString(36).substring(2, 8) +
+      Math.random().toString(36).substring(2, 8) +
+      ((Date.now() / 1000) | 0) +
+      "_" +
+      fileName
+    );
+  }
+
+  /**
+   * Devuelve el nombre del archivo a partir de un path o nombre del archivo.
+   *
+   * @param path Path o nombre del archivo.
+   * @param excludeExtension Si la extensión debe excluirse o no. Default false.
+   * @returns string con el nombre del archivo.
+   */
+  private fileNameFromPath(path: string, excludeExtension = false): string {
+    const slashIndex = path.lastIndexOf("/");
+    const slash = slashIndex !== undefined ? slashIndex + 1 : 0;
+
+    if (!excludeExtension) {
+      return path.substring(slash);
+    }
+
+    const dot = path.lastIndexOf(".");
+    return path.substring(slash, dot);
   }
 }