+++ src/app/state/user.state.ts
@@ -21,4 +21,9 @@ export interface UserState {
    * Current authenticated user.
    */
   user: WtUser | null;
+
+  /**
+   * Path to user's photo in localstorage.
+   */
+  userPhotoPath: string | null;
 }