+++ src/app/state/user.store.ts
@@ -18,6 +18,7 @@ export class UserStore extends BaseStore<UserState> {
       company: null,
       license: null,
       user: null,
+      userPhotoPath: null,
     });
   }
 }