+++ src/app/types/image-extension.type.ts
@@ -1,0 +1,2 @@
+/** Allowed extension for images. Android and IOS only support jpeg. */
+export type ImageExtension = "jpeg";